% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.fmmstil.K.parallel.random.R
\name{cluster.fmmstil.K.parallel.random}
\alias{cluster.fmmstil.K.parallel.random}
\title{Clustering using mstil given a fixed number of clusters in parallel using random initial parameters.}
\usage{
cluster.fmmstil.K.parallel.random(
  x,
  K,
  ncore = 1,
  numTrial = 1,
  criteria = c("ICL", "BIC", "AIC"),
  init.cluster.method,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{K}{a positive integer, represents the number of clusters in the model.}

\item{ncore}{a positive integer, represents the number of cpu threads to be used in parallel. By default 1.}

\item{numTrial}{a positive integer, represents the number of trials to be evaluated. By default 1.}

\item{criteria}{Either 'ICL', 'BIC', or 'AIC'. Represents the type of information criteria used for model selection. By default 'ICL'.}

\item{init.cluster.method}{a function of x, K that seperates x into K initial clusters.}

\item{init.param.method}{a function of x, returns initial parameters.}

\item{show.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, it accepts all control arguments used in fit.fmmstil.r and fit.fmmsil.}
}
\value{
a list with components:
\item{restricted}{a list containing details of the best fitted fmmstil.r.}
\item{unrestricted}{a list containing details of the best fitted fmmstil.}
\item{recordR}{a list of list containing details all fitted fmmstil.r.}
}
\description{
Clustering using mstil given a fixed number of clusters in parallel using random initial parameters.
}
\examples{
# Not run:
# data(RiverFlow)
# cluster.fmmstil.K.parallel.random(as.matrix(log(RiverFlow)),2,2)
}
