% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{cluster.fmmstil.K.parallel.windows}
\alias{cluster.fmmstil.K.parallel.windows}
\title{Clustering using mstil given a fixed number of clusters in parallel for windows.}
\usage{
cluster.fmmstil.K.parallel.windows(
  x,
  K,
  ncore = 1,
  numTrial = 1,
  init.cluster.method,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{K}{positive integer, number of cluster.}

\item{ncore}{number of cpu core to be used in parallel. By default 1.}

\item{numTrial}{a positive integer, number of trials to be evaluated.}

\item{init.cluster.method}{a function of x, K that seperates x into K initial clusters.}

\item{init.param.method}{a functino of x, return initial parameters.}

\item{show.progress}{show progress on console.}

\item{control}{list of control variables, it accepts all control arguments used in fit.fmmstil.r and fit.fmmsil. In this case, the default lambdaPenalty is 0.01 and the default cvgConR is 0.1 instead.}
}
\value{
a list with components:
\item{restricted}{a list containing details of the best fitted fmmstil.r.}
\item{unrestricted}{a list containing details of the best fitted fmmstil.}
\item{recordR}{a list of list containing details all fitted fmmstil.r.}
}
\description{
Clustering using mstil given a fixed number of clusters in parallel for windows.
}
\examples{
# Not run:
# data(RiverFlow)
# cluster.fmmstil.L.parallel(as.matrix(log(RiverFlow)),2,2)
}
