% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{cluster.fmmstil.K}
\alias{cluster.fmmstil.K}
\title{Clustering using mstil given a fixed number of clusters.}
\usage{
cluster.fmmstil.K(
  x,
  K,
  nstart,
  init.cluster.method,
  init.method,
  print.progress = TRUE,
  ...
)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{K}{positive integer, number of cluster.}

\item{nstart}{a positive integer, number of starting points to be evaluated.}

\item{init.cluster.method}{a function of x, K seperate x into initial clusters.}

\item{init.method}{a functino of x, return initial parameters.}

\item{print.progress}{show progress on console.}

\item{...}{additional parameters}
}
\value{
a list with components:
\item{Restricted}{a list containing details of the best fitted mstil.r.}
\item{Unrestricted}{a list containing details of the best fitted mstil.}
\item{record}{a list containing all fitted mstil.r.}
}
\description{
Clustering using mstil given a fixed number of clusters.
}
\examples{
# Not run:
}
