% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.fmmstil.parallel.random.R
\name{cluster.fmmstil.parallel.random}
\alias{cluster.fmmstil.parallel.random}
\title{Automatic model based clustering via fmmstil in parallel using random initial parameters.}
\usage{
cluster.fmmstil.parallel.random(
  x,
  ncore = 1,
  numTrial.fun,
  criteria = c("ICL", "BIC", "AIC"),
  init.cluster.method,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{ncore}{a positive integer, represents the number of cpu threads to be used in parallel. By default 1.}

\item{numTrial.fun}{a function of K that returns the number of trials to be evaluated.}

\item{criteria}{Either 'ICL', 'BIC', or 'AIC'. Represents the type of information criteria used for model selection. By default 'ICL'.}

\item{init.cluster.method}{a function of x, K that seperates x into K initial clusters.}

\item{init.param.method}{a function of x, returns initial parameters.}

\item{show.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, it accepts all control arguments used in fit.fmmstil.r and fit.fmmsil.}
}
\value{
a list with components:
\item{res}{a list containing details of the best fitted distribution.}
\item{record}{a list of lists containing details all fitted fmmstil.r.}
}
\description{
Automatic model based clustering via fmmstil in parallel using random initial parameters.
}
