% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{cluster.fmmstil.parallel}
\alias{cluster.fmmstil.parallel}
\title{Clustering using mstil given a fixed number of clusters.}
\usage{
cluster.fmmstil.parallel(
  x,
  ncore,
  nstart.fun,
  init.cluster.method,
  init.method,
  print.progress = TRUE,
  ...
)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{ncore}{a positive integer, number of core to be used}

\item{nstart.fun}{a function that returns nstart given K.}

\item{init.cluster.method}{a function of x, K seperate x into initial clusters.}

\item{init.method}{a functino of x, return initial parameters.}

\item{print.progress}{show progress on console.}

\item{...}{additional parameters}
}
\value{
a list with components:
\item{Res}{a list containing details of the best fitted distribution}
\item{Unrestricted}{a list containing details of the best fitted mstil.}
\item{record}{a list containing all fitted mstil.r for each K.}
}
\description{
Clustering using mstil given a fixed number of clusters.
}
\examples{
# Not run:
}
