% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{cluster.fmmstil}
\alias{cluster.fmmstil}
\title{Automatic model based clustering via fmmstil.}
\usage{
cluster.fmmstil(
  x,
  numTrial.fun,
  init.cluster.method,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{numTrial.fun}{a function of K that returns the number of trials to be evaluated.}

\item{init.cluster.method}{a function of x, K that seperates x into K initial clusters.}

\item{init.param.method}{a functino of x, return initial parameters.}

\item{show.progress}{show progress on console.}

\item{control}{list of control variables, it accepts all control arguments used in fit.fmmstil.r and fit.fmmsil. In this case, the default lambdaPenalty is 0.01 and the default cvgConR is 0.1 instead.}
}
\value{
a list with components:
\item{res}{a list containing details of the best fitted distribution.}
\item{record}{a list of list containing details all fitted fmmstil.r.}
}
\description{
Automatic model based clustering via fmmstil.
}
\examples{
# Not run:
# data(RiverFlow)
# cluster.fmmstil(as.matrix(log(RiverFlow)))
}
