% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmstil.R
\name{dmstil}
\alias{dmstil}
\title{This function return the quasi density function for mstil.}
\usage{
dmstil(x, lambda, delta, Ainv, nu, u, log.p = FALSE, control = list())
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{lambda}{skewing matrix with k rows.}

\item{delta}{location vector of size k.}

\item{Ainv}{lower triangular matrix, where t(Ainv) * Ainv is the precision matrix.}

\item{nu}{degree of freedom (>0).}

\item{u}{(Optional) a m x k matrix, a set of samples generated from standard k-dimensional multivariate t distribution with degree of freedom nu.}

\item{log.p}{a logical value. If TRUE, return the probability density function in logarithmic scale. By default FALSE.}

\item{control}{a list of control variables, see 'details'.}
}
\value{
return a numeric vector of length n.
}
\description{
This function return the quasi density function for mstil.
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{numLikSample}{number of samples used to estimate the density and log-likelihood functions. By default 1e6.}
}
}
