% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fmmstil.parallel.R
\name{fit.fmmstil.parallel}
\alias{fit.fmmstil.parallel}
\title{This function finds the penalised quasi likelihood estiamtes for finite mixture of mstil via EM in parallel.}
\usage{
fit.fmmstil.parallel(
  x,
  K,
  ncore = 1,
  param = NULL,
  init.cluster,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{K}{number of clusters.}

\item{ncore}{number of core to be used.}

\item{param}{list of lists of inital parameters, containing omega, lambda, delta, Ainv, and nu.}

\item{init.cluster}{(optional) initial clusters used to find initial parameters.}

\item{init.param.method}{(optional) method to obtain initial parameters. It needs to be a function of x and K, and return a list of list of parameters.}

\item{show.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{a vector of the estimated log-likelihood after each itereation.}
\item{par}{a list of lists of lists of fitted parameters after each iteration.}
\item{time}{a vector recorded the time elapsed after each iteration.}
}
\description{
This function finds the penalised quasi likelihood estiamtes for finite mixture of mstil via EM in parallel.
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{numLikSample}{a positive integer, represents the number of samples used to estimate the density and log-likelihood functions. By default 1e6. }
\item{conLevel}{a value between 0.5 and 1, represents the confidence level of the log-likelihood to be calculated. By default 0.95.}
\item{cvgN}{a positive integer. The algorithm stops when the estimated log-likelihood is not improved in cvgN iterations. By default 5.}
\item{lambdaPenalty}{a positive value, represents the L2 penalty coefficient for lambda. By default 1e-6.}
\item{ainvPenalty}{a positive value, represents the L2 penalty coefficient for Ainv. By default 1e-6.}
\item{maxit}{a positive integer, represents the maximum number of EM iterations allowed. By default 1e3.}
\item{maxitOptim}{a positive integer, represents the maximum number of iterations in optim allowed within each M-step. By default 10.}
\item{numGradSample}{a positive integer, represents the number of samples used to estimate the gradient. By default 1e4.}
\item{finDiffStep}{a positive value, represents the step size to estimate gradient w.r.t. nu. By default 1e-5.}
\item{stepSgd}{a positive value, represents the step size to be used in the stochastic gradient step to optimise nu. By default 1e-2.}
\item{iterSgd}{a positive integer, represents the number of iterations to be used in the stochastic gradient step to optimise nu. By default 1e2.}
\item{dimRateSgd}{a positive value, represents the diminishing rate for step size to be used in the stochastic gradient step to optimise nu. By default 1e-2.}
\item{batchSize}{a positive integer, represents the batch sample size. By default n.}
}
}
\examples{
# Not run:
# data(RiverFlow)
# fit.fmmstil(as.matrix(log(RiverFlow)), 2)
}
