% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fit.fmmstil.r}
\alias{fit.fmmstil.r}
\title{This function finds the maximum likelihood estiamtes for finite mixture of restricted mstil via EM.}
\usage{
fit.fmmstil.r(
  x,
  K,
  param,
  init.cluster,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{K}{number of clusters.}

\item{param}{list of lists of inital parameters, containing omega, lambda, delta, Ainv, and nu.}

\item{init.cluster}{(optional) initial clusters used to find initial parameters.}

\item{init.param.method}{(optional) method to obtain initial parameters. It needs to be a function of x that returns a list of list of parameters.}

\item{show.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{a vector recording the estimated log-likelihood after each itereation.}
\item{par}{a list of list of list of fitted parameters after each iteration.}
\item{time}{a vectorrecording the time elapsed after each iteration.}
}
\description{
This function finds the maximum likelihood estiamtes for finite mixture of restricted mstil via EM.
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{cvgConR}{tolerance level. The algorithm stops when increase of log-likelihood drops below this level. }
\item{lambdaPenalty}{L1 penalty coefficient for lambda. By default 0.}
\item{maxitR}{maximum number of EM iterations. By default 1000.}
\item{maxitOptimR}{maximum number of iterations in optim within each M-step. By default 1e3.}
}
}
\examples{
# Not run:
# data(RiverFlow)
# fit.fmmstil.r(as.matrix(log(RiverFlow)), 2)
}
