% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mstil.r.batch.R
\name{fit.mstil.r.batch}
\alias{fit.mstil.r.batch}
\title{This function finds the maximum penalised quasi likelihood estiamtes for mstil.r using batch of subsamples.}
\usage{
fit.mstil.r.batch(x, param, show.progress = TRUE, control = list())
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{param}{list of inital parameters, contains lambda, delta, Ainv, and nu.}

\item{show.progress}{logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{a vector of values of the log-likelihood function after each itereation.}
\item{par}{a list of lists of fitted parameters after each iteration.}
\item{time}{a vector recorded the time elapsed after each iteration.}
}
\description{
This function finds the maximum penalised quasi likelihood estiamtes for mstil.r using batch of subsamples.
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{cvgNR}{a positive integer. The algorithm stops when the estimated log-likelihood is not improved by at least cvgTolR in cvgNR iterations. By default 5.}
\item{cvgTolR}{a positive value. The algorithm stops when the estimated log-likelihood is not improved by at least cvgTolR in cvgNR iterations. By default 1e-2.}
\item{lambdaPenalty}{a positive value, represents the L1 penalty coefficient for lambda. By default 0.}
\item{ainvPenalty}{a positive value, represents the L2 penalty coefficient for Ainv. By default 0.}
\item{maxitR}{a positive integer, represents the maximum number iterations allowed. By default 1e3.}
\item{maxitOptimR}{a positive integer, represents the maximum number of iterations allowed in optim. By default 10.}
\item{batchSizeR}{a positive integer, represents the batch sample size. By default n.}
}
}
\examples{
# Not run:
# data(RiverFlow)
# fit.mstil.r.batch(as.matrix(log(RiverFlow)), control = list(batchSizeR = 100))
}
