% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fit.mstil.r}
\alias{fit.mstil.r}
\title{This function finds the maximum likelihood estimations for restricted mstil.}
\usage{
fit.mstil.r(x, param, control = list())
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{param}{list of inital parameters, containing lambda, delta, Ainv, and nu.}

\item{control}{list of control variables, see 'details'.
#@details The control argument is a list that accepts the following components.
\describe{
\item{lambdaPenalty}{L1 penalty coefficient for lambda. By default 0.}
\item{maxitOptimR}{maximum number of iterations in optim.. By default 1e3.}
}}
}
\value{
a list with components:
\item{lambda}{fitted lambda.}
\item{delta}{fitted delta.}
\item{Ainv}{fitted Ainv.}
\item{nu}{fitted nu.}
\item{logLik}{value of fitted log-Likelihood.}
}
\description{
This function finds the maximum likelihood estimations for restricted mstil.
}
\examples{
# Not run:
# data(RiverFlow)
# fit.mstil.r(as.matrix(log(RiverFlow)))
}
