% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fit.mstil}
\alias{fit.mstil}
\title{This function finds the maximum penalised quasi likelihood estiamtes for mstil.}
\usage{
fit.mstil(x, param, show.progress = TRUE, control = list())
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{param}{list of inital parameters, containing lambda, delta, Ainv, and nu.}

\item{show.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{a vector of the estimated log-likelihood function after each itereation.}
\item{par}{a list of list of fitted parameters after each iteration. Within each list, "lambda" is the fitted skewing matrix, "delta" is the location parameter, "Ainv" is the reparameterised scale parameter, "nu" is the degree of freedom.}
\item{logLikLower}{a numeric vector of the lower bound of the estimated log-likelihood function after each iteration.}
\item{logLikUpper}{a numeric vector of the upper bound of the estimated log-likelihood function after each iteration.}
\item{time}{a non-negative numeric vector, records the time elapsed after each iteration.}
}
\description{
This function finds the maximum penalised quasi likelihood estiamtes for mstil.
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{numLikSample}{number of samples used to estimate the density and log-likelihood functions. By default 1e6. }
\item{conLevel}{numeric between 0.5 and 1. Confidence level of the log-likelihood. By default 0.95.}
\item{cvgCon}{non-negative interger. The algorithm stops when the estimated maximum penalised log-likelihood is not improved in cvgCon iterations. By default 5.}
\item{lambdaPenalty}{L1 penalty coefficient for lambda. By default 0.}
\item{maxit}{maximum number iterations. By default 1000.}
\item{maxitOptim}{maximum number of iterations in optim. By default 10.}
\item{numGradSample}{number of samples used to estimate the gradient. By default 1e4.}
\item{finDiffStep}{step size to estimate gradient w.r.t. nu. By default 1e-5.}
\item{stepSgd}{step size to be used in the stochastic gradient step. By default 1e-2.}
\item{iterSgd}{number of iterations to be used in the stochastic gradient step. By default 1e2.}
\item{dimRateSgd}{dimishing rate for step size to be used in the stochastic gradient step. By default 1e-2.}
}
}
\examples{
# Not run:
# data(RiverFlow)
# fit.mstil(as.matrix(log(RiverFlow)))
}
