% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fmmstil.fitness}
\alias{fmmstil.fitness}
\title{This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil}
\usage{
fmmstil.fitness(x, param, u, control = list())
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{param}{list of lists of fitted parameters, containing omega, lambda, delta, Ainv, and nu.}

\item{u}{(Optional) a m x k matrix, a set of samples generated from standard k-dimensional multivariate t distribution with degree of freedom nu.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{the estimated log likelihood.}
\item{ICL}{the estiamted integrated completed log likelihood.}
\item{AIC}{the estiamted Akaike information criteria.}
\item{BIC}{the estiamted Bayesian information criteria.}
\item{clust}{the estimated group membership.}
}
\description{
This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{numLikSample}{number of samples used to estimate the density and log-likelihood functions. By default 1e6.}
}
}
