% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fmmstil.r.fitness}
\alias{fmmstil.r.fitness}
\title{This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil.r}
\usage{
fmmstil.r.fitness(x, param)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{param}{list of lists of fitted parameters, containing omega, lambda, delta, Ainv, and nu.}
}
\value{
a list with components:
\item{logLik}{the log likelihood.}
\item{ICL}{the integrated completed log likelihood.}
\item{AIC}{the Akaike information criteria.}
\item{BIC}{the Bayesian information criteria.}
\item{clust}{the estimated group membership.}
}
\description{
This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil.r
}
