% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{mstil.logL}
\alias{mstil.logL}
\title{This function estimate the value of log likelihood function}
\usage{
mstil.logL(x, lambda, delta, Ainv, nu, u, sample.mc = 10000, q = 0.95)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{lambda}{a numeric skewing matrix of size k x p.}

\item{delta}{a numeric vector of length k, representing the location parameter.}

\item{Ainv}{an lower triangular numeric matrix of size k x k, where t(Ainv) "times" Ainv representing the inverse of the scale parameter.}

\item{nu}{a positive value representing the degree of freedom/}

\item{u}{(optional) a m x k matrix, representing samples generated by a standard k-dimensional multivariate t distribution with degree of freedom nu.}

\item{sample.mc}{(optional) a positive integer. If u is not specified, generate "sample.mc" samples from a standard k-dimensional multivariate t distribution with degree of freedom nu. By default 10000.}

\item{q}{a value in range (0.5,1), representing the 2-sided confidence interval percentage. By default 0.95.}
}
\value{
a list with components:
\item{logL.lower}{the lower bound of the estimated log-likelihood function.}
\item{logL}{the estimate of the log-likelihood function.}
\item{logL.upper}{the upper bound of the estimated log-likelihood function.}
}
\description{
This function estimate the value of log likelihood function
}
\examples{
# Not run:
# data(RiverFlow)
# k <- ncol(RiverFlow)
# lambda <- diag(k)
# delta <- rep(0,k)
# Ainv <- diag(k)
# nu <- 2
# mstil.logL(log(RiverFlow), lambda, delta, Ainv, nu)
}
