% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmstil.r.R
\name{dmstil.r}
\alias{dmstil.r}
\title{This function returns to likelihood function of restricted mstil.}
\usage{
dmstil.r(x, lambda, delta, Ainv, nu, log.p = FALSE)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{lambda}{skewing matrix with k rows.}

\item{delta}{location vector of size k.}

\item{Ainv}{lower triangular matrix, where t(Ainv) * Ainv is the precision matrix.}

\item{nu}{degree of freedom (>0).}

\item{log.p}{a logical value. If TRUE, return the probability density function in logarithmic scale. By default FALSE.}
}
\value{
return a numeric vector of length n.
}
\description{
This function returns to likelihood function of restricted mstil.
}
\examples{
# Not run:
# data(RiverFlow)
# k <- ncol(RiverFlow)
# lambda <- diag(k)
# delta <- rep(0,k)
# Ainv <- diag(k)
# nu <- 2
# dmstil.r(as.matrix(log(RiverFlow)), lambda, delta, Ainv, nu)
}
