% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{dmstil}
\alias{dmstil}
\title{This function return the quasi density function for mstil.}
\usage{
dmstil(x, lambda, delta, Ainv, nu, u, sample.mc = 10000, log.p = FALSE)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{lambda}{a numeric skewing matrix of size k x p.}

\item{delta}{a numeric vector of length k, representing the location parameter.}

\item{Ainv}{an upper triangular numeric matrix of size k x k, where t(Ainv) "times" Ainv representing the inverse of the scale parameter.}

\item{nu}{a positive value representing the degree of freedom.}

\item{u}{(Optional) a m x k matrix, representing samples generated by a standard k-dimensional multivariate t distribution with degree of freedom nu.}

\item{sample.mc}{(Optional) a positive integer. If u is not specified, generate "sample.mc" samples from a standard k-dimensional multivariate t distribution with degree of freedom nu. By default 10000.}

\item{log.p}{a logical value. If TRUE, return the probability density function in logarithmic scale. By default FALSE.}
}
\value{
gives a numeric vector of length n.
}
\description{
This function return the quasi density function for mstil.
}
\examples{
# Not run:
data(RiverFlow)
k <- ncol(RiverFlow)
lambda <- diag(k)
delta <- rep(0,k)
Ainv <- diag(k)
nu <- 2
dmstil(log(RiverFlow), lambda, delta, Ainv, nu)
}
