% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fit.fmmstil.r}
\alias{fit.fmmstil.r}
\title{This function finds the maximum likelihood estiamtes for finite mixture of restricted mstil via EM.}
\usage{
fit.fmmstil.r(
  x,
  K,
  omega,
  lambda,
  delta,
  Ainv,
  nu,
  maxit.bfgs = 1000,
  maxit = 1000,
  convergence.tol = 0.01,
  print.progress = TRUE
)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{K}{a positive integer, number of clusters.}

\item{omega}{a positive numeric vector of length K, where sum of the vector equals to 1.}

\item{lambda}{list of K numeric skewing matrices of size k x p.}

\item{delta}{list of K numeric vectors of length k, representing the location parameter.}

\item{Ainv}{list of lower triangular numeric matrices of size k x k, where t(Ainv) "times" Ainv representing the inverse of the scale parameter.}

\item{nu}{list of positive values representing the degree of freedom.}

\item{maxit.bfgs}{a positive integer, representing the maximum number of iterations used in the MLE step. By default 1000.}

\item{maxit}{a positive integer, representing the maximum number of iterations. By default 1000.}

\item{convergence.tol}{the maximum tolerance before the algorithm stops.}

\item{print.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}
}
\value{
a list with components:
\item{logL}{a vector of the estimated log-likelihood function after each itereation.}
\item{par}{a list of list of fitted parameters after each iteration. Within each list, "lambda" is the fitted skewing matrix, "delta" is the location parameter, "Ainv" is the reparameterised scale parameter, "nu" is the degree of freedom.}
\item{time}{a non-negative numeric vector, records the time elapsed after each iteration.}
}
\description{
This function finds the maximum likelihood estiamtes for finite mixture of restricted mstil via EM.
}
\examples{
# Not run:
}
