% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.fmmstil.r.R
\name{fit.fmmstil.r}
\alias{fit.fmmstil.r}
\title{This function finds the maximum likelihood estiamtes for finite mixture of restricted mstil via EM.}
\usage{
fit.fmmstil.r(
  x,
  K,
  param,
  init.cluster,
  init.param.method,
  show.progress = TRUE,
  control = list()
)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{K}{number of clusters.}

\item{param}{list of lists of inital parameters, containing omega, lambda, delta, Ainv, and nu.}

\item{init.cluster}{(optional) initial clusters used to find initial parameters.}

\item{init.param.method}{(optional) method to obtain initial parameters. It needs to be a function of x that returns a list of list of parameters.}

\item{show.progress}{a logical value. If TRUE, progress of the algorithm will be printed in console. By default TRUE.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{a vector of the estimated log-likelihood after each itereation.}
\item{par}{a list of lists of lists of fitted parameters after each iteration.}
\item{time}{a vector recorded the time elapsed after each iteration.}
}
\description{
This function finds the maximum likelihood estiamtes for finite mixture of restricted mstil via EM.
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{cvgNR}{a positive integer. The algorithm stops when the estimated log-likelihood is not improved by at least cvgTolR in cvgNR iterations. By default 5.}
\item{cvgTolR}{a positive value. The algorithm stops when the estimated log-likelihood is not improved by at least cvgTolR in cvgNR iterations. By default 1e-2.}
\item{lambdaPenalty}{a positive value, represents the L1 penalty coefficient for lambda. By default 0.}
\item{maxitR}{a positive integer, represents the maximum number of EM iterations allowed. By default 1000.}
\item{maxitOptimR}{a positive integer, represents the maximum number of iterations in optim allowed within each M-step. By default 1e2.}
\item{batchSizeR}{a positive integer, represents the batch sample size. By default n.}
}
}
\examples{
# Not run:
# data(RiverFlow)
# fit.fmmstil.r(as.matrix(log(RiverFlow)), 2)
}
