% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mstil.R
\name{fit.mstil.r}
\alias{fit.mstil.r}
\title{This function finds the maximum likelihood estimations for restricted mstil.}
\usage{
fit.mstil.r(x, lambda, delta, Ainv, nu, maxit = 1000, max.lambda = Inf)
}
\arguments{
\item{x}{a n x k matrix, representing n k-variate samples.}

\item{lambda}{a numeric diagonal skewing matrix of size k x k.}

\item{delta}{a numeric vector of length k, representing the location parameter.}

\item{Ainv}{an lower triangular numeric matrix of size k x k, where t(Ainv) "times" Ainv representing the inverse of the scale parameter.}

\item{nu}{a positive value representing the degree of freedom/}

\item{maxit}{a positive integer, representing the maximum number of iterations. By default 1000.}

\item{max.lambda}{positive value, representign the maximum lambda accepted.}
}
\value{
a list with components:
\item{logL}{a vector of the estimated log-likelihood function after each itereation.}
\item{par}{a list of list of fitted parameters after each iteration. Within each list, "lambda" is the fitted skewing matrix, "delta" is the location parameter, "Ainv" is the reparameterised scale parameter, "nu" is the degree of freedom.}
\item{time}{a non-negative numeric vector, records the time elapsed after each iteration.}
}
\description{
This function finds the maximum likelihood estimations for restricted mstil.
}
\examples{
# Not run:
}
