% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mstil.r.R
\name{fit.mstil.r}
\alias{fit.mstil.r}
\title{This function finds the maximum likelihood estimations for restricted mstil.}
\usage{
fit.mstil.r(x, param, control = list())
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{param}{list of inital parameters, contains lambda, delta, Ainv, and nu.}

\item{control}{list of control variables, see 'details'.
#@details The control argument is a list that accepts the following components.
\describe{
\item{lambdaPenalty}{a positive value, represents the L2 penalty coefficient for lambda. By default 0.}
\item{ainvPenalty}{a positive value, represents the L2 penalty coefficient for Ainv. By default 0.}
\item{maxitOptimR}{a positive integer, represents the maximum number of iterations allowed in optim. By default 1e3.}
}}
}
\value{
a list with components:
\item{lambda}{the value of fitted lambda.}
\item{delta}{the value of fitted delta.}
\item{Ainv}{the value of fitted Ainv.}
\item{nu}{the value of fitted nu.}
\item{logLik}{the value of fitted log-Likelihood.}
}
\description{
This function finds the maximum likelihood estimations for restricted mstil.
}
\examples{
# Not run:
# data(RiverFlow)
# fit.mstil.r(as.matrix(log(RiverFlow)))
}
