% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmstil.fitness.R
\name{fmmstil.fitness}
\alias{fmmstil.fitness}
\title{This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil}
\usage{
fmmstil.fitness(x, param, u, control = list())
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{param}{list of lists of parameters, contains list of omega, list of lambda, list delta, list of Ainv, and list of nu.}

\item{u}{(Optional) list of K m x k matrices, each matrix contain samples generated from standard k-dimensional multivariate t distribution with degree of freedom of the K-th cluster.}

\item{control}{list of control variables, see 'details'.}
}
\value{
a list with components:
\item{logLik}{the value of log-likelihood.}
\item{ICL}{the value of integrated completed log-likelihood.}
\item{AIC}{the value of Akaike information criterion.}
\item{BIC}{the value of Bayesian information criterion.}
\item{clust}{the estimated group membership.}
}
\description{
This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil
}
\details{
The control argument is a list that accepts the following components.
\describe{
\item{numLikSample}{number of samples used to estimate the density and log-likelihood functions. By default 1e6.}
}
}
