% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmmstil.r.fitness.R
\name{fmmstil.r.fitness}
\alias{fmmstil.r.fitness}
\title{This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil.r}
\usage{
fmmstil.r.fitness(x, param)
}
\arguments{
\item{x}{matrix of quantiles of size n x k. Each row is taken as a quantile.}

\item{param}{list of lists of parameters, contains list of omega, list of lambda, list delta, list of Ainv, and list of nu.}
}
\value{
a list with components:
\item{logLik}{the value of log-likelihood.}
\item{ICL}{the value of integrated completed log-likelihood.}
\item{AIC}{the value of Akaike information criterion.}
\item{BIC}{the value of Bayesian information criterion.}
\item{clust}{the estimated group membership.}
}
\description{
This function returns the log likelihood, ICL, AIC, BIC, and group memberships for fmmstil.r
}
